♻️ Move tests into files they test
